requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
Flask-Login==0.6.3
flask-mail==0.10.0
flask-wtf==1.2.2
pyotp==2.9.0
qrcode==8.0
google-auth
google-auth-oauthlib
google-auth-httplib2
